+++ packages/vsce-api/src/interfaces/resources/IBundle.ts
@@ -17,4 +17,5 @@ export interface IBundle extends IResourceWithEnableStatus {
   enablestatus: string;
   name: string;
   partcount: string;
+  availstatus: string;
 }