+++ packages/vsce-api/src/interfaces/resources/IBundlePart.ts
@@ -15,4 +15,6 @@ export interface IBundlePart extends IResourceWithEnableStatus {
   bundle: string;
   bundlepart: string;
   partclass: string;
+  availstatus: string;
+  parttype: string;
 }