+++ packages/vsce-api/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe Explorer for IBM CICS Transaction Server API will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added attributes to resource interfaces: JVM Server, TS Queue, Bundle, Bundle Part, Library Dataset. [#573](https://github.com/zowe/cics-for-zowe-client/pull/573)
+
 ## `6.18.0`
 
 - Enhancement: Added IRegion, IManagedRegion interface. [#571](https://github.com/zowe/cics-for-zowe-client/pull/571)