+++ packages/vsce/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cics-extension-for-zowe" extension will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Updated highlight information for CICS Resources: JVM Server, TS Queue, Bundle, Bundle Part, JVM Endpoint, Library Dataset. [#573](https://github.com/zowe/cics-for-zowe-client/pull/573)
+
 ## `3.18.0`
 
 - BugFix: Fixed Resource Inspector action menu so that it does not display beyond the panel edges. [#572](https://github.com/zowe/cics-for-zowe-client/issues/572)